+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # v3.1.0
+* Changed: Properly throw NoSuchElementException from Iterables when next is called with no more elements
 * Added: Accumulo in table storage of streaming property value data
 * Added: Functions to delete extended data columns
 * Added: Multi-value extended data columns
 * Added: Sorting and aggregating of row ids and table names
 * Added: Option to execute a script from the shell
+* Fixed: Streaming property value input stream from data table
 * Removed: ElasticSearch 1.x support
 
 # v3.0.4